requirements.txt
selenium 
pymongo 
requests 
flask 
dnspython
python-dotenv
os